pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = ["Miguel Silva <71125410+msilva94@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "url_shortener"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-wtf = "^1.1.1"
flask-sqlalchemy = "^3.0.5"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
